I've fixed some unit issues with the interlacing model so now I can get to the observed G141 fluxes straight from the F140W direct fluxes without any fudge factors.
I also adjusted the zeroth order calibration and normalization for slightly better agreement with the observed spectrum.

The new file "intersim.py" is a place for running simulations based on observed pointings, sort of like aXeSIM.  The idea is that you take and observed direct image and can model line and continuum depths by taking artificial templates and "dispersing" them with the G141 calibration and the *observed* direct profiles.  This will then give you continuum/line depth directly as a function of object size, concentration, and mag, at least assuming that the input morphology distribution is representative.    

aXeSIM can sort of do these things, but I think in a more awkward way with input catalogs and thumbnails.  Note that this method also uses the FLT error extensions directly to generate noise.  And finally, you can fit the simulated spectrum, with errors, with all of the same tools that we've developed for the interlaced reduction (redshift, line flux, etc.).



git-svn-id: https://subversion.assembla.com/svn/threedhst_internal/trunk@688 f9184c78-529c-4a83-b317-4cf1064cc5e0