intersim: Improved plotting of simulation results.
reduce.deep_model: this script generates a model for every object in an interlaced exposure down to faint limits (M140=28) to evaluate the background subtraction of the grism exposures.  With the model, you can generate an aggressive mask and look at the pixel distribution.  On the first few pointings I checked, there was a slight oversubtraction of about 0.002-0.003 e/s (native pixels), corresponding to continuum at H~27-28.

git-svn-id: https://subversion.assembla.com/svn/threedhst_internal/trunk@702 f9184c78-529c-4a83-b317-4cf1064cc5e0