Added new functions to adjust how the line fluxes and equivalent widths are fit, where I now explicitly include a correction to scale the observed spectrum to the level of the photometry.  The output results are pretty much identical for the two methods, though the linefit.dat now lists a line like
# tilt correction to photometry: s0 = -0.239 [ 0.025 ], s1 = 0.106 [ 0.205 ]

showing the tilt correction that would be necessary and there is an additional column in the table, "scale_to_photom", provided that would be used to scale the measured line fluxes to the level defined by the photometry.  In most cases this factor is > 1, perhaps consistent with the expectation that the optimal spectral extraction is missing some flux with respect to the total photometric fluxes.

To use the new code, simply call "new_fit_free_emlines" rather than "fit_free_emlines".


git-svn-id: https://subversion.assembla.com/svn/threedhst_internal/trunk@740 f9184c78-529c-4a83-b317-4cf1064cc5e0